docker/service-fetcher/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.7.1
pymongo>=4.10,<5
openai>=0.27.0
faiss-cpu==1.8.0
numpy==1.26.4
python-dateutil>=2.8.2
httpx==0.27.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.7.1